+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2020-02-29
+### Added
+- Laravel 7 compatibility.
+
+### Fixed
+- asset insertion scripts.
+
 ## [0.10.10] - 2020-02-23
 ### Fixed
 - file input field rendering issue for tailwind template.